CSHARP-3181: Default DateTimeOffset serialization makes it impossible to search across timezones
