+++ tests/MongoDB.Bson.Tests/Serialization/Serializers/DateTimeOffsetSerializerTests.cs
@@ -14,10 +14,13 @@
 */
 
 using System;
+using System.Globalization;
 using System.IO;
+using System.Linq;
 using FluentAssertions;
 using MongoDB.Bson.IO;
 using MongoDB.Bson.Serialization;
+using MongoDB.Bson.Serialization.Attributes;
 using MongoDB.Bson.Serialization.Serializers;
 using MongoDB.TestHelpers.XunitExtensions;
 using Xunit;
@@ -27,16 +30,16 @@ namespace MongoDB.Bson.Tests.Serialization.Serializers
     public class DateTimeOffsetSerializerTests
     {
         [Fact]
-        public void constructor_with_no_arguments_should_return_expected_result()
+        public void Constructor_with_no_arguments_should_return_expected_result()
         {
             var subject = new DateTimeOffsetSerializer();
 
-            subject.Representation.Should().Be(BsonType.Array);
+            subject.Representation.Should().Be(BsonType.Document);
         }
 
         [Theory]
         [ParameterAttributeData]
-        public void constructor_with_representation_should_return_expected_result(
+        public void Constructor_with_representation_should_return_expected_result(
             [Values(BsonType.Array, BsonType.DateTime, BsonType.Document, BsonType.String)] BsonType representation)
         {
             var subject = new DateTimeOffsetSerializer(representation);
@@ -45,7 +48,7 @@ namespace MongoDB.Bson.Tests.Serialization.Serializers
         }
 
         [Fact]
-        public void constructor_with_representation_should_throw_when_representation_is_invalid()
+        public void Constructor_with_representation_should_throw_when_representation_is_invalid()
         {
             var exception = Record.Exception(() => new DateTimeOffsetSerializer(BsonType.Null));
 
@@ -109,6 +112,69 @@ namespace MongoDB.Bson.Tests.Serialization.Serializers
             }
 
             result.Should().Be(DateTimeOffset.Parse(expectedResult));
+        }
+
+        [Fact]
+        public void Equals_null_should_return_false()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+
+            var result = x.Equals(null);
+
+            result.Should().Be(false);
+        }
+
+        [Fact]
+        public void Equals_object_should_return_false()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+            var y = new object();
+
+            var result = x.Equals(y);
+
+            result.Should().Be(false);
+        }
+
+        [Fact]
+        public void Equals_self_should_return_true()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+
+            var result = x.Equals(x);
+
+            result.Should().Be(true);
+        }
+
+        [Fact]
+        public void Equals_with_equal_fields_should_return_true()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+            var y = new DateTimeOffsetSerializer(BsonType.Array);
+
+            var result = x.Equals(y);
+
+            result.Should().Be(true);
+        }
+
+        [Fact]
+        public void Equals_with_not_equal_field_should_return_false()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+            var y = new DateTimeOffsetSerializer(BsonType.DateTime);
+
+            var result = x.Equals(y);
+
+            result.Should().Be(false);
+        }
+
+        [Fact]
+        public void GetHashCode_should_return_zero()
+        {
+            var x = new DateTimeOffsetSerializer(BsonType.Array);
+
+            var result = x.GetHashCode();
+
+            result.Should().Be(0);
         }
 
         [Theory]
