+++ tests/MongoDB.Bson.Tests/Serialization/Serializers/NetPrimitiveSerializerTests.cs
@@ -24,6 +24,7 @@ using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Attributes;
 using MongoDB.Bson.Serialization.Options;
 using MongoDB.Bson.Serialization.Serializers;
+using MongoDB.TestHelpers.XunitExtensions;
 using Xunit;
 
 namespace MongoDB.Bson.Tests.Serialization
@@ -834,112 +835,6 @@ namespace MongoDB.Bson.Tests.Serialization
         public void GetHashCode_should_return_zero()
         {
             var x = new CultureInfoSerializer();
-
-            var result = x.GetHashCode();
-
-            result.Should().Be(0);
-        }
-    }
-
-    public class DateTimeOffsetSerializerTests
-    {
-        public class TestClass
-        {
-            [BsonRepresentation(BsonType.Array)]
-            public DateTimeOffset A { get; set; }
-            [BsonRepresentation(BsonType.Document)]
-            public DateTimeOffset D { get; set; }
-            [BsonRepresentation(BsonType.String)]
-            public DateTimeOffset S { get; set; }
-        }
-
-        // TODO: more DateTimeOffset tests
-
-        [Fact]
-        public void TestSerializeDateTimeOffset()
-        {
-            var value = new DateTimeOffset(new DateTime(2010, 10, 8, 11, 29, 0), TimeSpan.FromHours(-4));
-            var isoDate = value.UtcDateTime.ToString("yyyy-MM-ddTHH:mm:ss.FFFFFFFK", CultureInfo.InvariantCulture);
-            var obj = new TestClass
-            {
-                A = value,
-                D = value,
-                S = value
-            };
-            var json = obj.ToJson();
-            var expected = "{ 'A' : #A, 'D' : #D, 'S' : '#S' }";
-            expected = expected.Replace("#A", string.Format("[NumberLong('{0}'), {1}]", value.DateTime.Ticks, value.Offset.TotalMinutes));
-            expected = expected.Replace("#D",
-                "{ 'DateTime' : ISODate('#D'), 'Ticks' : NumberLong('#T'), 'Offset' : #O }"
-                    .Replace("#D", isoDate)
-                    .Replace("#T", value.DateTime.Ticks.ToString())
-                    .Replace("#O", JsonConvert.ToString(value.Offset.TotalMinutes)));
-            expected = expected.Replace("#S", "2010-10-08T11:29:00-04:00");
-            expected = expected.Replace("'", "\"");
-            Assert.Equal(expected, json);
-
-            var bson = obj.ToBson();
-            var rehydrated = BsonSerializer.Deserialize<TestClass>(bson);
-            Assert.True(bson.SequenceEqual(rehydrated.ToBson()));
-        }
-
-        [Fact]
-        public void Equals_null_should_return_false()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
-
-            var result = x.Equals(null);
-
-            result.Should().Be(false);
-        }
-
-        [Fact]
-        public void Equals_object_should_return_false()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
-            var y = new object();
-
-            var result = x.Equals(y);
-
-            result.Should().Be(false);
-        }
-
-        [Fact]
-        public void Equals_self_should_return_true()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
-
-            var result = x.Equals(x);
-
-            result.Should().Be(true);
-        }
-
-        [Fact]
-        public void Equals_with_equal_fields_should_return_true()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
-            var y = new DateTimeOffsetSerializer(BsonType.Array);
-
-            var result = x.Equals(y);
-
-            result.Should().Be(true);
-        }
-
-        [Fact]
-        public void Equals_with_not_equal_field_should_return_false()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
-            var y = new DateTimeOffsetSerializer(BsonType.DateTime);
-
-            var result = x.Equals(y);
-
-            result.Should().Be(false);
-        }
-
-        [Fact]
-        public void GetHashCode_should_return_zero()
-        {
-            var x = new DateTimeOffsetSerializer(BsonType.Array);
 
             var result = x.GetHashCode();
